Copyright Headers + test of ramdisk load
Ok, so a lot of fixes were neccessary, returning to 16 bit was a bit
more complicated than initial thought.

Anyway it seems to work now, it loads my kernel first, relocates it to
1mb mark, then returns to 16 bit mode, loads the ramdisk, goes 32 bit,
relocates it to the 2mb mark.

Next up is implementation of the new module system in MCore.